apps/backend/requirements.txt
Django==5.1.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.3.2
python-decouple==3.8
gunicorn==23.0.0
